+++ class/GaiaAlpha/Model/Menu.php
@@ -1,0 +1,8 @@
+<?php
+namespace GaiaAlpha\Model;
+
+class Menu extends BaseModel
+{
+    protected static $table = 'menus';
+    protected static $fillable = ['title', 'location', 'items'];
+}